+++ lib/screens/stepper_screen/question_screen/widgets/question_screen_buttons.dart
@@ -11,7 +11,7 @@ import 'package:provider/provider.dart';
 class QuestionScreenButtons extends StatefulWidget {
   final List<int> relationIdHistory;
   final List<QuestionModel> selectedQuestions;
-  final List<int> userInputs;
+  final List<double> userInputs;
   final QuestionBloc bloc;
   final int serviceId;
   final int selectedAnswerId;
@@ -63,7 +63,8 @@ class _QuestionScreenButtonsState extends State<QuestionScreenButtons> {
                   }
                 }
               } else if (lastQuestion.type == 'radio') {
-                provider.removeItem(price: lastQuestion.items[0].price);
+                provider.removeItem(
+                    price: lastQuestion.items[0].price.toDouble());
               }
               widget.selectedQuestions.removeLast();
               widget.bloc.add(
